Fix type definitions for uint32 and int32

+++ include/types.h
@@ -5,12 +5,12 @@
 #define FALSE 0
 
 typedef unsigned long long uint64;
-typedef unsigned long uint32;
+typedef unsigned int uint32;
 typedef unsigned short uint16;
 typedef unsigned char uint8;
 
 typedef long long int64;
-typedef long int32;
+typedef int int32;
 typedef short int16;
 typedef char int8;
